issue #42, fix use of uninitialized variables, fix invalid reads, fix compile warnings, fix memory leaks (at exit), added missing modes to luma magick fx

+++ veejay-current/veejay-server/libvjmsg/vj-common.c
@@ -154,7 +154,7 @@ void 	veejay_print_backtrace()
 void	veejay_print_backtrace()
 {
 	void *space[100];
-	size_t i,s;
+	int i,s;
 	char **strings;
 
 	s = backtrace( space, 100 );
